tests/wpt/tests/css/css-view-transitions/navigation/at-rule-opt-in-auto.html

<!DOCTYPE html>
<title>View Transitions: @view-transition opt in for auto</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/">
<link rel="author" href="mailto:bokan@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@view-transition {
  navigation: auto;
}
</style>
<script>
const params = new URLSearchParams(location.search);
const mode = params.get("mode") || "test";

switch (mode) {
case "test":
  promise_test(async t => {
    const result = await new Promise(resolve => {
      window.did_reveal = msg => { resolve(msg) };
      const popup = window.open("?mode=old");
      t.add_cleanup(() => popup.close());
    });
    assert_equals(result, "pass");
  });
  break;
case "old":
  onload = () => requestAnimationFrame(() => requestAnimationFrame(() => {
    location.replace("?mode=new");
  }));
  break;
case "new":
  onpagereveal = e => {
    window.opener.did_reveal(e.viewTransition ? "pass" : "fail");
  };
  break;
}
</script>
